+++ aten/src/ATen/ATen.h
@@ -10,9 +10,7 @@
 #include <ATen/DynamicLibrary.h>
 #include <ATen/Formatting.h>
 #include <ATen/Functions.h>
-#ifdef BUILD_NAMEDTENSOR
 #include <ATen/NamedTensor.h>
-#endif
 #include <ATen/ScalarOps.h>
 #include <ATen/Tensor.h>
 #include <ATen/TensorGeometry.h>
