+++ aten/src/ATen/native/Distributions.cpp
@@ -12,9 +12,7 @@
 #include <ATen/native/Distributions.h>
 #include <ATen/native/DispatchStub.h>
 #include <ATen/native/UnaryOps.h>
-#ifdef BUILD_NAMEDTENSOR
 #include <ATen/NamedTensorUtils.h>
-#endif
 
 #include <type_traits>
 #include <functional>
