+++ aten/src/ATen/native/cuda/CUDAUnaryOps.cpp
@@ -1,8 +1,6 @@
 #include <ATen/ATen.h>
 #include <ATen/LegacyTHFunctionsCUDA.h>
-#ifdef BUILD_NAMEDTENSOR
 #include <ATen/NamedTensorUtils.h>
-#endif
 
 namespace at { namespace native {
