+++ aten/src/ATen/templates/LegacyTHFunctions.cpp
@@ -4,9 +4,7 @@
 
 #include <ATen/ATen.h>
 #include <ATen/Utils.h>
-#ifdef BUILD_NAMEDTENSOR
 #include <ATen/NamedTensorUtils.h>
-#endif
 #include <ATen/${Generator}.h>
 #include <ATen/ExpandUtils.h>
 ${th_headers}
