+++ aten/src/ATen/test/NamedTensor_test.cpp
@@ -1,4 +1,3 @@
-#ifdef BUILD_NAMEDTENSOR
 #include <gtest/gtest.h>
 
 #include <ATen/ATen.h>
@@ -6,6 +5,7 @@
 #include <c10/util/Exception.h>
 #include <c10/util/C++17.h>
 
+#ifdef BUILD_NAMEDTENSOR
 using at::Dimname;
 using at::DimnameList;
 using at::NamedTensorMeta;
