+++ aten/src/THC/generic/THCTensorMathBlas.cu
@@ -3,9 +3,7 @@
 #else
 
 #include "ATen/cuda/CUDAContext.h"
-#ifdef BUILD_NAMEDTENSOR
 #include <ATen/NamedTensorUtils.h>
-#endif
 
 #define ERROR_ONLY_FP_TYPES(func) \
   THError("%s for CUDA tensors only supports floating-point types. Try converting the tensors with .float()", func);
