+++ torch/csrc/python_dimname.cpp
@@ -1,9 +1,9 @@
-#ifdef BUILD_NAMEDTENSOR
 #include <torch/csrc/python_dimname.h>
 #include <torch/csrc/Exceptions.h>
 #include <torch/csrc/utils/python_strings.h>
 #include <c10/util/flat_hash_map.h>
 
+#ifdef BUILD_NAMEDTENSOR
 namespace torch {
 
 struct InternedStringsTable {
