+++ torch/csrc/python_dimname.h
@@ -1,8 +1,8 @@
 #pragma once
-#ifdef BUILD_NAMEDTENSOR
 #include <torch/csrc/python_headers.h>
 #include <ATen/Dimname.h>
 
+#ifdef BUILD_NAMEDTENSOR
 at::Dimname THPDimname_parse(PyObject* obj);
 bool THPUtils_checkDimname(PyObject* obj);
 bool THPUtils_checkDimnameList(PyObject* obj);
